+++ src/main/java/net/commoble/looot/enchantment/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package net.commoble.looot.enchantment;
+
+import org.jspecify.annotations.NullMarked;