+++ src/main/java/net/commoble/looot/loot/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package net.commoble.looot.loot;
+
+import org.jspecify.annotations.NullMarked;